pyproject.toml
[tool.poetry]
name = "django-learning-system"
version = "0.1.0"
description = ""
authors = ["Nikita Solouyou <solovev.nikita.05@gmail.com>"]
readme = "README.md"
packages = [{include = "django_learning_system"}]

[tool.poetry.dependencies]
python = "^3.11"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
ruff = "^0.0.290"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
